requirements.txt
pynput
psutil
Pillow
pyperclip
pandas
matplotlib